// management/frontend/src/api/schemas/physicalKitPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */
import type { KitDeviceSlot } from './kitDeviceSlot';

export interface PhysicalKitPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 128
     */
  serial_number: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  name: string;
  template_id: string;
  /** @maxLength 128 */
  terminal_serial?: string;
  /** @maxLength 128 */
  bound_username?: string;
  /** @maxLength 32 */
  status?: string;
  /** @maxLength 256 */
  location?: string;
  /** @maxLength 1000 */
  remark?: string;
  workspace_id: string;
  template_name?: string;
  device_slots?: KitDeviceSlot[];
  creator_name?: string | null;
  updater_name?: string | null;
}